Reorganize text models
