+++ fastai/text/models/__init__.py
@@ -1,0 +1,2 @@
+from .awd_lstm import *
+__all__ = [*awd_lstm.__all__]
