+++ fastai/text/models/qrnn.py
